Added feature: Display random popular movie with similar movies
This commit implements a feature to display a random popular movie along with its similar movies. It fetches popular movies data using the useGetPopularMovie hook and selects a random movie from the fetched data. Then, it fetches similar movies for the selected movie using the useMutation hook. Once the data is loaded, it displays the selected popular movie with its details and a list of similar movies. The list is limited to a maximum of 10 similar movies.