Added display of recent movies
Implemented the display of recent movies on the homepage. Fetches recent movie data using the useGetRecentMovie hook with the current year. Displays loading skeletons for movie cards while data is being fetched. Once the data is loaded, renders up to 20 recent movies with their posters and titles in a scrollable container.